requirements.txt
requests==2.25.0
PyYAML==5.4.1